PYR-686: Add "frosted glass" look to all map tools
Find a styling method for applying a "frosted glass" treatment to the
backgrounds of all map tool controls. Isolate and move inline background
properties applied to tool controls, and cull them into a new styling
function `tool-background`: that can be composed into the the `tool`
styling function and freely applied to the Collapsible Panel.