PYR-719: Add conditional tab visibility
The `get-org-list` api request takes an optional user-id argument to filter
organizations to those that a user belongs to.

A tab-restriction property was added to the "NV Energy" tab in
`config.cljs`

This is picked up while mapping over the list of potential tabs. If a
particular tab has a organization restriction and if that organization
is not in the list returned by `get-org-list`, the tab visibility is set
to hidden.