Add funcool/decimal library use to-significant fn
The change, replaces an application of `u/precision` with an application
of a fun and cool `to-significant` function. The first argument to the function, the
precision, is set to a single digit of significance and the second
argument is left out: the default :round-half-up is what we want.

TODO: This looks great for the small values, 10e-5, desired by
plignrate, but some testing is needed, since a bug is probably being
introduced here for values greater than 10, where we would want more
than a single digit of significance.