Add underlay tooltip note about lack of point info
Add a helpful note, on the tooltip for the Optional Layers, that highlights
that underlays do not have point information that can be queried. It'll
serve as a guiding note for users that presume the Point Information
tool may provide data when clicking on a visible region of an underlay.