Re-add `call-sql-async` for further work on admin
While reviewing pull-request
https://github.com/pyregence/pyregence/pull/705 with Gary, we
identified that further work to simply adding and editing users was
needed on Pyrecast. Absence of the `call-sql-async` function
implemntation was noted and copied over again in utils.cljs file.
The function was recommended as a tool that could be used to simplify
interaction with the database server to create a richer set of
functionality and UX to empower admins with more intuitive user management.