Fix "Reactive deref not supported in lazy seq..." warning.
I guess this is something to be aware of when attempting to deref an atom in a lazy sequence. The containing component
won't immediately evaluate the sequence and so changes to the drefed'd atom would go unnoticed. To fix, either wrap lazy
sequence producing expression in a `doall` or some code rearrangement to dereference the atom elsewhere.

https://github.com/reagent-project/reagent/issues/18