explore two options for showing km and mi.
The "show both option" shows the user both km and mi at the same time.
The "radio button option" shows the user only one or the other and gives them
a choice to toggle between.

Note that the radio button option has some style choices to clean up, but
were not going that route atm so it's being kept around mostly for notes.