PYR1-970 add miles and kilometers to measure distance tool modal (#890)
* explore two options for showing km and mi.

The "show both option" shows the user both km and mi at the same time.
The "radio button option" shows the user only one or the other and gives them
a choice to toggle between.

Note that the radio button option has some style choices to clean up, but
were not going that route atm so it's being kept around mostly for notes.

* remove "toggle between units" option

remove the toggle between units option as that's not the direction
the feature is heading.

* clarify the conversion of the distance units

by bringing the relevant information to the top and closer together.

* minor style and html changes

I'm using the html P tag instead of label because labels are slightly more for connecting to html
input tags. Also, i removed the padding because I couldn't notice the difference
and wasn't sure it was relevant.

* Draw the users eye to the distance more

The distance is what the user is here for, it should stand out a bit.

Also this commit has some clean up. I movde where distance was in html tree
because it didn't need to be nested under that flex div, that was for the
buttons.

* add a comment about why were doing this.

This comment might not survive the PR.

* parens convey non critical information

and a forward slash can be mistaken as implying division.

* remove comment as the commit will be enough

* clarify code by writing out the steps

This is slightly better because it isolates the conversion such that a reader
can, at a glance, understand what were going to and from.

* remove unused require

* unify conversion functionality around multiplication.

* remove extra space

* sort hashmap by keys

* align columns

* move logic to top level let block.

We can't bind miles (mi) or kilometers (km) in a reagent with-let
because that only, as per the docs, only is evaluated/bound once.

So instead we bind some helper functions.

* prefer shorter lines