remove inactive cameras
Inactive means cameras older then four hours. Because the wildfire non-California
api doesn't indicate the timezone of when their cameras last refreshed, we have
have to reformat them so they have a utc timezone so they can, along side the
California cameras (which are marked as utc), be removed if they are inactive.

Is it safe to assume the non-cali cameras have a utc timestamp? good question.
For one, I believe it's what we were doing before, so were no less safe.
Secondly, I spot checked a couple cameras that were very up-to-date (within an hour)
and those times seemed to correspond to UTC time.
That fact, along with the knowledge that the cali cameras do use utc, gives
me high confidence that they are currently unmarked utc timestamps.

Which is important to get right, as the margin for error can be quite large if
the api, is using some other timezone (likely the local timezone of the camera
if anything).

However, we should set a filter to catch changes, log and alert, as the api
might change from under us at any time.

Also note that the arrangement of this code currently favors a tighter scope,
it might be ideal though, to make this code more accessible because much of it
deals with handling the api. That could mean it's own namespace, or even it's
own library: a wildfire api library.