Improve researchability
Improve researchability by exposing the individual parts of the `get-cameras`
function. This allows someone to do research on the wildfire api and makes sure
it's responding with the data we need, and that were correctly using that data.

This is exactly what we do at the end of this file in the `scratch zone` where
we explore how many cameras are inactive. And we look into how many wildfire
cameras don't have timestamps that are marked with a timezone, and see that we
can correctly assign them one.

Notice that the scratch zone code doesn't need to make it into the main branch.
I personally find it useful to keep some code and data like this ling around in
a namespace, others prefer it goes in a test or goes away. I'm flexible.

Regardless, its possible this overall effort of isolating out the functions
should go even further, and we should create a wildfire api library where such
functions and live and be used by others. But we can and should delay that
choice until we know more about the requirements.