add selecting default Risk Ignition Pattern through option directly
This change adds the ability to select a default Risk Ignition Pattern
(e.g UtilityCompany Energy Lines) by updating the database's

Organization Layers table layer_config column to have contain a string
that's a valid edn with a top level key :selected? set to true. e.g

layer_config                                                                           |
---------------------------------------------------------------------------------------|
{:selected? true :opt-label "UtilityCompany", :filter "...", :geoserver-key "..."}|

This works because that selected key will be available in the panel_drowdown
because it's assoced into state along the Organization Layer tabels layer_path. e.g

layer_path                                          |
----------------------------------------------------|
[:fire-risk :params :pattern :options :UtilityCompany]|

And this is available on a per company selection

|org_layer_uid|organization_rid|
|-------------|----------------|
|98           |20              |

The major downside here is it introduces yet another way to select
the default, it takes precedence over the :default-option found in the config.cljs
which is used as the "val" in the panel_dropdown.

Why not use the :default-option? Because that will require a change to how we
collect the state of the layer_config. Will explore that option in another commit.