PYR1-1014 remove camera icons with inactive feeds as icons on the map (#891)
* remove inactive cameras

Inactive means cameras older then four hours. Because the wildfire non-California
api doesn't indicate the timezone of when their cameras last refreshed, we have
have to reformat them so they have a utc timezone so they can, along side the
California cameras (which are marked as utc), be removed if they are inactive.

Is it safe to assume the non-cali cameras have a utc timestamp? good question.
For one, I believe it's what we were doing before, so were no less safe.
Secondly, I spot checked a couple cameras that were very up-to-date (within an hour)
and those times seemed to correspond to UTC time.
That fact, along with the knowledge that the cali cameras do use utc, gives
me high confidence that they are currently unmarked utc timestamps.

Which is important to get right, as the margin for error can be quite large if
the api, is using some other timezone (likely the local timezone of the camera
if anything).

However, we should set a filter to catch changes, log and alert, as the api
might change from under us at any time.

Also note that the arrangement of this code currently favors a tighter scope,
it might be ideal though, to make this code more accessible because much of it
deals with handling the api. That could mean it's own namespace, or even it's
own library: a wildfire api library.

* Improve researchability

Improve researchability by exposing the individual parts of the `get-cameras`
function. This allows someone to do research on the wildfire api and makes sure
it's responding with the data we need, and that were correctly using that data.

This is exactly what we do at the end of this file in the `scratch zone` where
we explore how many cameras are inactive. And we look into how many wildfire
cameras don't have timestamps that are marked with a timezone, and see that we
can correctly assign them one.

Notice that the scratch zone code doesn't need to make it into the main branch.
I personally find it useful to keep some code and data like this ling around in
a namespace, others prefer it goes in a test or goes away. I'm flexible.

Regardless, its possible this overall effort of isolating out the functions
should go even further, and we should create a wildfire api library where such
functions and live and be used by others. But we can and should delay that
choice until we know more about the requirements.

* remove scratch space

* emphasize that binding is unused by following common

clojure convention of underscoring it

* remove frontend logic that deals with the cameras being too old

Remove the frontend logic to check if the cameras are too old, because it will
never get used because the backend is only sending the frontend cameras that are
up to date.

Put another way, the backend now does what the frontend was doing here.

* add needed logic to handle the image-src

* move comments closer to where they are relevant

* remove unused imports

* grammar improvements