PYR1-1016 logged in users map should stay where they are when switching between tabs (#893)
* Disable auto zoom for PSPS tab.

* disable auto zoom on ignition pattern

To enable the auto zoom for a particular ignition pattern, for instance, a
utility companies ignition pattern, then configure it in the pyregence postgres
database like so:

INSERT INTO public.organization_layers (organization_rid,layer_path,layer_config) VALUES
   (6,'[:fire-risk :params :pattern :options :some-utility-co]','{:auto-zoom? true}');

Notice the rest of the keys in the auto-zoom? map are omitted.

* Reinstating auto zoom when non utility ignition pattern is selected on the risk tab

---------

Co-authored-by: Oliver-BE <obaldwinedwards@gmail.com>