WIP Add toast message when there are no active fires.
This change adds a toast message when there are no active fires
by explicitly checking if the tab selected is active fires and
then relying on processed-params to contain information about all active fires
intended to be used in the layer selection fire name drop down, which includes
"all active fires" which is why we decrement by one.

The implicate nature of using whats meant to be drop down information for the count
isn't ideal and should be improved, but this commit at least proves the concept.

Relying on the `key` being equal to `active-fire` seems fine on reflection,
as the intent is to show the toast when the tab is clicked.

Even beyond the possible complications mentioned above, this is a WIP because
it doesn't fulfill the ticket because this toast will happen every time the tab
is clicked, and not just the first time.