WIP Show 'no active fires' based on explicit state
Before this change, showing the 'no active fires' toast was based on the
process-params :options state, which is intended to hold the dropdown options
for "fire names". This change improves on that by instead explicitly relying on
state `show-no-active-fires?` thats gathered nearer to the source (the api call).

This branch is still a WIP because it doesn't just show the toast
'the first time' the user clicks the 'active fires' tab as desired, and instead,
and will show it every time.