Toast "No active fires." when no active fires and first tab visit
This is achieved by keeping track of the number of times the active fires tab
has been clicked.

Coordinating the active fire count and the active fire tab click count,
towards deciding if the toast should be shown,
can be done a number of ways, this commit offers one option that
unifies around checking of both counts are still zero.