Weather and Risk Tab Forecast Start time is now always UTC.
This is achieved by giving each location the time is shown the option to
interpret the time zone slider request to update that location. For example,
on the weather tab, the forecast start time option receives the request to change
the timezone from :utc to :local and always returns :utc.

This change also lets each location set a different default to accommodate
the feature request in 1026 that Weather and Risk Forecast Start Time always
be UTC.