set timezones per model parmas
This seems ok given other obvious options, we didnt want it on the model-init,
and if it goes alongside the top level key (e.g fire-weather) its not easily
accessible on the params where the time data is stored. Also, setting
along side the time-slider? key would imply it might impact the time-slider,
which it shouldn't.