PYR1-975 add toast message when there are no active fires (#896)
* WIP Add toast message when there are no active fires.

This change adds a toast message when there are no active fires
by explicitly checking if the tab selected is active fires and
then relying on processed-params to contain information about all active fires
intended to be used in the layer selection fire name drop down, which includes
"all active fires" which is why we decrement by one.

The implicate nature of using whats meant to be drop down information for the count
isn't ideal and should be improved, but this commit at least proves the concept.

Relying on the `key` being equal to `active-fire` seems fine on reflection,
as the intent is to show the toast when the tab is clicked.

Even beyond the possible complications mentioned above, this is a WIP because
it doesn't fulfill the ticket because this toast will happen every time the tab
is clicked, and not just the first time.

* WIP Show 'no active fires' based on explicit state

Before this change, showing the 'no active fires' toast was based on the
process-params :options state, which is intended to hold the dropdown options
for "fire names". This change improves on that by instead explicitly relying on
state `show-no-active-fires?` thats gathered nearer to the source (the api call).

This branch is still a WIP because it doesn't just show the toast
'the first time' the user clicks the 'active fires' tab as desired, and instead,
and will show it every time.

* minor grammar change

* Toast "No active fires." when no active fires and first tab visit

This is achieved by keeping track of the number of times the active fires tab
has been clicked.

Coordinating the active fire count and the active fire tab click count,
towards deciding if the toast should be shown,
can be done a number of ways, this commit offers one option that
unifies around checking of both counts are still zero.

* fix grammar

* Show toast everytime layer is fetch, which will happen everytime

the tab is clicked with how things are currently setup. I'm not sure
if fetching layers happens outside clicking the tab.

* only show layers toast if no-fires toast can't happen.

* Update src/cljs/pyregence/pages/near_term_forecast.cljs

Co-authored-by: Oliver Baldwin Edwards <obaldwinedwards@gmail.com>

* empty calls seq, no need to call it twice

* align let binding values.

* remove redundant toast message

This should have been removed in the commit where the functionality was added
into the cond block.

* Small updates

---------

Co-authored-by: Oliver Baldwin Edwards <obaldwinedwards@gmail.com>