remove per model default timezone and update risk tab timezone fn
The risk tab/forecast should always be utc. Given the described
behavior, all the tabs/forecasts can share the global timezone as the default.
But it's nice to leave a space for a default in the hashmap. If we do
want a per tab/forecast timezone, then another piece of state will need
to be created as the order will be selected-timezone (originally set to nil),
tab/forecast default, set in the config map, followed by a global timezone state
which could be set at the top level of the config namespace as a var.