PYR1-1030 Select the forecast after the mapbox loads. (#906)
* Select the forecast after the mapbox loads.

Selecting the forecast should be done after the mapbox loads because
Mapbox asynchronously fetches styles, and likely does other work asynchronously in it's
constructor, which means the app needs a way to do work after that async work finishes.

That work cannot be done before, as doing it before will cause issues like
the active fire icons not loading, which is really more like a symptom of the
whole mapbox program failing.

To facility doing the work explicitly after the mapbox loads, we take advantage
of the Mapbox's lifecycle load event. Here is the link:

https://docs.mapbox.com/mapbox-gl-js/api/map/#map.event:load

and here is an example from that link:

```
// Initialize the map
const map = new mapboxgl.Map({});
// Set an event listener that fires
// when the map has finished loading.
map.on('load', () => {
    console.log('A load event occurred.');
});
```

To mirror this example in Pyregence, this commit passes the `on-load-fn` to select the forecast
to the mapbox js object called `the-map` after it's returned out of the ratom
reset!

I'll note that I'm not entirely sure why the mapbox object is inside a ratom in
the first place, as it's already global mutable state. Is it used to somehow
update facilitate rendering a component?

* reset the map only after it loads.

This would seem to change what a watcher or validator on the atom would see.
It's slightly more correct that the atom should only hold a loaded map at this
stage in the app execution.

* No need for a to pull from the go channel

... returned by `select-forecast` because it's just used as a side effect.
This change isn't directly related to the desire to improve the loading,
it's just a clean up.