WIP fix mapbox height loading issue
The issue with the mapbox height seems related to setting that height in the
root-component via javascript, which can sometimes, somehow, get stuck at a
value smaller then the full viewport.

So, rather then have javascript try to figure out the viewport, I figured we can
try setting the height to the full viewport and simultaneously make sure the
mapbox is resized after it's loaded, as opposed to when the component mounted,
which would seem an unrelated event in my mind (which is why i assume it was
resizing based on a timeout).

For reasons I don't fully understand, setting the height to be the full
viewport, seems to necessitate fiddling with the absolute position bottoms a bit
to bring things up higher.

It also means I don't think we need much of whats going on in this update-fn in
the root-components component-did-mount function, which all seems aimed at
trying to resize the parent element of the canvas which should now just fill the
entire viewport.

I would say I have medium level confidence this is an improvement. I haven't
fully vetted this change on mobile or tried to understand the update-fn.