Improve Mapbox resizing
This commit improves mapbox resizing story by calling `mapbox.resize` after
mapbox loads, as opposed to on a timer which could be wrong, and by allowing the
mapbox's html tags parent height to be 100% instead of calculating it via the
innerHeight.

My gut says the less Javascript involved in timing and styling the better off
the browser will behave in general, and testing so far seems to show parity.
However, as it's hard to test the actual problem this commit is meant to
address (when the map partially loads) however, the jury is still out on how
useful this change is overall.

Setting the height to 100% in the html tree necessitated some changes to the
collapsible panel tree such that the "help-section" could maintain it's position
at the bottom of the page.