WIP Users should only be able to see their organizations information.
The key principle here is that the request handler needs to make a choice given
both information the user can set and information they can't. In this case,
that's the user-id from the params and session. The handler will ignore the
params because the user shouldn't be able to see another organizations
information.

Or maybe they should if their an admin, in which case the handler should lookup,
based on their user id information in the session, if there an admin, and
possible allow the request.