A user should only be able to get their own email.
This is a hard story for me to piece together, why is a user trying to get their
own email? What role is this supposed to be? `token` isn't a very good way
to communicate roles, a token is supposed to be a collection of information
that computer exchange, we need a much more pyregence specific term like
pyregence.role/user, or just user would be fine to atm.