PYR1-1038/Improve Mapbox resizing (#919)
* Improve Mapbox resizing

This commit improves mapbox resizing story by calling `mapbox.resize` after
mapbox loads, as opposed to on a timer which could be wrong, and by allowing the
mapbox's html tags parent height to be 100% instead of calculating it via the
innerHeight.

My gut says the less Javascript involved in timing and styling the better off
the browser will behave in general, and testing so far seems to show parity.
However, as it's hard to test the actual problem this commit is meant to
address (when the map partially loads) however, the jury is still out on how
useful this change is overall.

Setting the height to 100% in the html tree necessitated some changes to the
collapsible panel tree such that the "help-section" could maintain it's position
at the bottom of the page.

* scroll no longer flows into layer selection header

* return scroll bar to terms of use page

I think this was hiding a scroll bar that another change caused to no longer
appear.

* use react fragment to group when styling isn't needed

* the resize causes the flash and it might not be needed

I'm not sure how to re-create the original issue so its possible simply not
dynamically setting the height fixes it and the resize on load isn't needed.

* allow link at the bottom to be part of overflow

* keep refresh button anchored to top of the page.