wip building ubjar
this commit is mostly to checkpoint the work at a point where the jar builds
such that it can run the server.

for that to work local/root triangulum needs to point
at the branch that moves the manifest to the classpath.

clearly the hardcoded output-to value is suspect, but i'm not sure what the
manifest is really doing with that key value pair so i'm not sure what the more
'elegant solution is'. its a pain to have to update that value and understand
the ramifications of the change and i'm not sure how the manifest.edn is being
used. e.g why is it a key value when the key is equal to the value?

Theoretically this could be used for fingerprinting, but is it? if not, it seems
strange to use the manifest to communicate the location of the app.js back to
triangulum, which again, already has very strict rules about where it should be.