move the cache refresh logic forward
I looked at the clojure library chime and talked to people on the clojure
slack channel:

https://clojurians.slack.com/archives/C03S1KBA2/p1755765659400009

they seem to think the loop would be just fine as long as
i didn't let the exceptions stop the outer loop from trying again.

The reality is though if we get one exception, chances are the whole pull
is off and ill have to come fix it.