Improve how we show latest observation.
unicode > text, and the alt-label for h_1 is weird. We could do more here, but
this seems to fix the current units were "seeing". I didn't sample and that wouldn't
help as observations could send different units at anytime, so having the wmo
is a good fallback.

We could look into using the wmo notation label instead of the alt-label for some.
might be overkill.