Correctly use react components.
In general we need to call [cmpt, not (cmpt, e.g not (wrap-page
or (wrap-component. Because [cmpt allows for reagent to enact the react
lifecycle, e.g passing it's props to `:reagent-render`.

This change, from ( to [, triggers an issue because,
some root-component, wrapped over `params` (should be called args) and
added the `forecast-type`. So to re-produce that, we, as it idiomatic in clojure,
bundled the root-component and the forecast-type into a hashmap and moved data
in and out of it.

This isn't pretty, I believe whats out of sync here is that there are really
only to components that need this, short-term-forecast and long-term-forecast.

So then, i think we should just make those component, and those stateful
components, should be aware of their stateful differences.

Basically, i propose that react state is  best composed through react
components, not through clojure closures.