Use form 3 component (#992)
* Add blank account settings page.

* Add a mock nav bar.

* Wire together existing /account-settings components.

* Set settings page to use font family Roboto

* use longer name bc react was complaining.

* remove empty tabs that are conditioned out.

* Fetch organizations and conditional change nav bar tabs.

* Expose user name on settings page.

By returning it in the verify_user_login sql fn which has comment that implies
it should be returning the name anyway, and then passing it to the client session.

* Small alterations to pass name on.

* prevent box from looking to long.

* no longer need name-part if using full name.

* Clicking log out now actually logs you out!

* commented out because component isn't ready.

* Remove old comment.

* format role type.

* Hopefully improved grammar.

* Use full name.

* refactor.

* remove unused code.

* Route the user name from the db to the settings page.

As the TODO notes, we copied a function from admin we should share or consolidate.
Also, it seems like sql `verify_user_email` and `set_user_password` implicitly
share a return table, is there a better way to sync their return values?

* Only call the backend if we don't have orgs.

The other code formed a loop because it kept re-triggering the backend call.

* Starting state needs to be nil to trigger initial call.

* Add padding to search icon.

* Add TODO about adding on-click event.

* comment out info tab until we have associated text.

* Consider having a namespace that handles all things roles.

* Let button grow taller on collapse.

* Give some space between elements.

* The set height seems more wrong then right.

* use correct role name and add auth notes.

* Only account manager and super admin can see unaffiliated members tab.

* turn into a function so it's easier to test.

* make fn private because it probably wont be used elsewhere...

* improve doc and clean up.

* formatting.

* Clean up, clarify, and grammar improvements.

* The `Roboto` font-family cascades down from the page settings.

And we also remove some defaults.

* quick fixes.

* WIP this doesnt render the orgs, their fetched, but the component

isn't updated.

* print orgs.

* wip debug notes.

* I believe we need to call [wrap-page rather then (wrap-page...

* Correctly use react components.

In general we need to call [cmpt, not (cmpt, e.g not (wrap-page
or (wrap-component. Because [cmpt allows for reagent to enact the react
lifecycle, e.g passing it's props to `:reagent-render`.

This change, from ( to [, triggers an issue because,
some root-component, wrapped over `params` (should be called args) and
added the `forecast-type`. So to re-produce that, we, as it idiomatic in clojure,
bundled the root-component and the forecast-type into a hashmap and moved data
in and out of it.

This isn't pretty, I believe whats out of sync here is that there are really
only to components that need this, short-term-forecast and long-term-forecast.

So then, i think we should just make those component, and those stateful
components, should be aware of their stateful differences.

Basically, i propose that react state is  best composed through react
components, not through clojure closures.

* Reorganize the routing info to have less indirection and bindings.